rename: api > supabase.ts를 api > supabaseCSR > supabase.ts로 파일 위치를 변경하였습니다.
