+++ src/components/my-owl/meeting/Meeting.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { getUserId } from '@/utils/my-owl/getUserId';
-import { getRoomData, getUsersData, getRoomParticipantsId, getUserMeetingsId } from '@/api/supabase';
+import { getRoomData, getUsersData, getRoomParticipantsId, getUserMeetingsId } from '@/api/supabaseCSR/supabase';
 import { sortMeetingInfo } from '@/utils/my-owl/meeting/sortMeetingInfo';
 
 import styles from './Meeting.module.css';
